tickets/src/routes/show.ts
import express, { Request, Response } from "express";
import { Ticket } from "../models/ticket";
import { NotFoundError } from "@odticketing/common";
import mongoose from "mongoose";

const router = express.Router();

router.get("/api/tickets/:id", async (req: Request, res: Response) => {
  const isValidId = mongoose.Types.ObjectId.isValid(req.params.id);
  if (!isValidId) {
    throw new NotFoundError();
  }
  const ticket = await Ticket.findById(req.params.id);
  if (!ticket) {
    throw new NotFoundError();
  }
  res.send(ticket);
});

export { router as showTicketRouter };
